requirements.txt
Django~=4.2.7
django-seed~=0.3.1
psycopg2-binary~=2.9.9
django-location-field~=2.7.2
django-debug-toolbar~=4.2.0
django-bootstrap-v5~=1.0.11 
django-bootstrap-icons~=0.8.6
django-jquery~=3.1.0
django-bootstrap-datepicker-plus~=5.0.4
django-staticfiles-moment~=2.0.0
django-crispy-forms~=2.0
crispy-bootstrap5~=2023.10
django-leaflet
djangorestframework~=3.14.0
#GDAL
#pysqlite3
